chore(arkmarket): move utils to proper dir
