+++ apps/arkmarket/src/components/token/token-actions-create-listing.tsx
@@ -44,8 +44,8 @@ import type { Token } from "~/types";
 import durations from "~/constants/durations";
 import { env } from "~/env";
 import usePrices from "~/hooks/usePrices";
-import formatAmount from "~/lib/formatAmount";
 import getCollection from "~/queries/getCollection";
+import formatAmount from "~/utils/formatAmount";
 import ToastExecutedTransactionContent from "./toast-executed-transaction-content";
 import ToastRejectedTransactionContent from "./toast-rejected-transaction-content";
 import TokenActionsTokenOverview from "./token-actions-token-overview";
