+++ apps/arkmarket/src/components/token/token-actions-make-offer.tsx
@@ -44,7 +44,7 @@ import { ETH } from "~/constants/tokens";
 import { env } from "~/env";
 import useBalance from "~/hooks/useBalance";
 import useConnectWallet from "~/hooks/useConnectWallet";
-import formatAmount from "~/lib/formatAmount";
+import formatAmount from "~/utils/formatAmount";
 import ToastExecutedTransactionContent from "./toast-executed-transaction-content";
 import ToastRejectedTransactionContent from "./toast-rejected-transaction-content";
 import TokenActionsTokenOverview from "./token-actions-token-overview";
