+++ apps/arkmarket/src/components/token/token-offers-table-row.tsx
@@ -11,7 +11,7 @@ import { TableCell, TableRow } from "@ark-market/ui/table";
 
 import type { Token, TokenMarketData, TokenOffer } from "~/types";
 import ConnectWalletModal from "~/components/connect-wallet-modal";
-import ownerOrShortAddress from "~/lib/ownerOrShortAddress";
+import ownerOrShortAddress from "~/utils/ownerOrShortAddress";
 import AcceptOffer from "./accept-offer";
 import CancelOffer from "./cancel-offer";
