+++ apps/arkmarket/src/components/wallet/portfolio.tsx
@@ -8,11 +8,11 @@ import { validateAndParseAddress } from "starknet";
 
 import type { ViewType } from "~/components/view-type-toggle-group";
 import type { WalletTokensApiResponse } from "~/queries/getWalletData";
+import { getWalletTokens } from "~/queries/getWalletData";
 import {
   walletCollectionFilterKey,
   walletCollectionFilterParser,
-} from "~/lib/wallet-search-params";
-import { getWalletTokens } from "~/queries/getWalletData";
+} from "~/utils/walletSearchParams";
 import PortfolioActivityData from "./portfolio-activity-data";
 import PortfolioHeader from "./portfolio-header";
 import PortfolioItemsData from "./portfolio-items-data";
