+++ apps/arkmarket/src/utils/formatAmount.ts
