+++ apps/arkmarket/src/utils/ownerOrShortAddress.ts
