+++ apps/arkmarket/src/utils/sortTraitsByCountAsc.ts
