+++ apps/arkmarket/src/utils/walletSearchParams.ts
